refactored for separate beamline dev
